pyproject.toml
[tool.poetry]
name = "rhizk-api"
version = "0.1.0"
description = "api for rhizk"
authors = ["jaikumarpr"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
pydantic = "^2.8.2"
openai = "^1.40.3"
numpy = "^2.0.1"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
scipy = "^1.14.0"
pymemcache = "^4.0.0"
nanoid = "^2.0.0"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
httpx = "^0.27.0"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
